+++ dot/parachain/overseer/overseer.go
@@ -11,10 +11,10 @@ import (
 
 	availability_store "github.com/ChainSafe/gossamer/dot/parachain/availability-store"
 	"github.com/ChainSafe/gossamer/dot/parachain/backing"
+	"github.com/ChainSafe/gossamer/dot/parachain/chainapi"
 	collatorprotocolmessages "github.com/ChainSafe/gossamer/dot/parachain/collator-protocol/messages"
 	parachain "github.com/ChainSafe/gossamer/dot/parachain/runtime"
 	parachaintypes "github.com/ChainSafe/gossamer/dot/parachain/types"
-	"github.com/ChainSafe/gossamer/dot/parachain/util"
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/common"
@@ -137,7 +137,7 @@ func (o *Overseer) processMessages() {
 
 				subsystem = o.nameToSubsystem[parachaintypes.AvailabilityStore]
 
-			case util.ChainAPIMessage[util.Ancestors], util.ChainAPIMessage[util.BlockHeader]:
+			case chainapi.ChainAPIMessage[chainapi.Ancestors], chainapi.ChainAPIMessage[chainapi.BlockHeader]:
 				subsystem = o.nameToSubsystem[parachaintypes.ChainAPI]
 
 			case parachain.RuntimeAPIMessage:
